WIP: Refer to templates directly (by file name)
This change removes the "template group" file and instead makes it so
that each template is stored in its own file. Saved template instances
refer to the template by the file name.

It is an attempt at simplifying the system, which may be easier than
filling in the remaining loose ends with the template group files and
could also have a better usability.